refactor speech command example
