+++ example/speech_command/.gitignore
@@ -1,0 +1,2 @@
+data/
+models/
