+++ example/speech_command/Makefile
@@ -1,0 +1,5 @@
+.POHNY: train
+train:
+	mkdir -p models data
+	python3 sc/train.py
+
