--- example/speech_command/code/dataset.py
@@ -1,63 +1,0 @@
-import torch
-from torchaudio.datasets import SPEECHCOMMANDS
-import os
-
-
-def pad_sequence(batch):
-    # Make all tensor in a batch the same length by padding with zeros
-    batch = [item.t() for item in batch]
-    batch = torch.nn.utils.rnn.pad_sequence(batch, batch_first=True,
-                                            padding_value=0.)
-    return batch.permute(0, 2, 1)
-
-
-def label_to_index_fn(labels):
-    return lambda word: torch.tensor(labels.index(word))
-
-
-def index_to_label_fn(labels):
-    return lambda index: labels[index]
-
-
-def collate_fn_lambda(labels):
-    label_to_index = label_to_index_fn(labels)
-    return lambda batch: collate_fn(batch, label_to_index)
-
-
-def collate_fn(batch, label_to_index):
-    # A data tuple has the form:
-    # waveform, sample_rate, label, speaker_id, utterance_number
-
-    tensors, targets = [], []
-
-    # Gather in lists, and encode labels as indices
-    for waveform, _, label, *_ in batch:
-        tensors += [waveform]
-        targets += [label_to_index(label)]
-
-    # Group the list of tensors into a batched tensor
-    tensors = pad_sequence(tensors)
-    targets = torch.stack(targets)
-
-    return tensors, targets
-
-
-class SubsetSC(SPEECHCOMMANDS):
-    def __init__(self, subset: str = None):
-        super().__init__("../data/", download=True)
-
-        def load_list(filename):
-            filepath = os.path.join(self._path, filename)
-            with open(filepath) as fileobj:
-                return [os.path.normpath(os.path.join(self._path, line.strip()))
-                        for line in fileobj]
-
-        if subset == "validation":
-            self._walker = load_list("validation_list.txt")
-        elif subset == "testing":
-            self._walker = load_list("testing_list.txt")
-        elif subset == "training":
-            excludes = load_list("validation_list.txt") + load_list(
-                "testing_list.txt")
-            excludes = set(excludes)
-            self._walker = [w for w in self._walker if w not in excludes]
