--- example/speech_command/requirements.txt
@@ -1,4 +1,0 @@
-torch
-torchaudio
-numpy
-tqdm==4.64.0
