+++ example/speech_command/sc/model.py
@@ -14,9 +14,6 @@ class M5(nn.Module):
         self.conv3 = nn.Conv1d(n_channel, 2 * n_channel, kernel_size=3)
         self.bn3 = nn.BatchNorm1d(2 * n_channel)
         self.pool3 = nn.MaxPool1d(4)
-        self.conv4 = nn.Conv1d(2 * n_channel, 2 * n_channel, kernel_size=3)
-        self.bn4 = nn.BatchNorm1d(2 * n_channel)
-        self.pool4 = nn.MaxPool1d(4)
         self.fc1 = nn.Linear(2 * n_channel, n_output)
 
     def forward(self, x):
@@ -29,9 +26,6 @@ class M5(nn.Module):
         x = self.conv3(x)
         x = F.relu(self.bn3(x))
         x = self.pool3(x)
-        x = self.conv4(x)
-        x = F.relu(self.bn4(x))
-        x = self.pool4(x)
         x = F.avg_pool1d(x, x.shape[-1])
         x = x.permute(0, 2, 1)
         x = self.fc1(x)
